requirements.txt
fastapi>=0.120.4
uvicorn[standard]>=0.38.0
pydantic>=2.12.3
boto3>=1.40.64
transformers>=4.57.1
torch>=2.9.0
accelerate>=1.11.0
python-dotenv>=1.2.1